refactor(EditProfileActivity): Create FileUtils.kt to save images to internal storage
